use case insensitive headers in all events
