+++ aws-lambda-java-events/src/main/java/com/amazonaws/services/lambda/runtime/events/APIGatewayProxyRequestEvent.java
@@ -1,5 +1,7 @@
 package com.amazonaws.services.lambda.runtime.events;
 
+import com.amazonaws.services.lambda.runtime.events.models.HttpHeaders;
+
 import java.io.Serializable;
 import java.util.*;
 
@@ -16,9 +18,9 @@ public class APIGatewayProxyRequestEvent implements Serializable, Cloneable {
 
     private String httpMethod;
 
-    private Headers<String> headers = new Headers<>();
+    private HttpHeaders<String> headers;
 
-    private Headers<List<String>> multiValueHeaders = new Headers<>();
+    private HttpHeaders<List<String>> multiValueHeaders;
 
     private Map<String, String> queryStringParameters;
 
@@ -33,77 +35,6 @@ public class APIGatewayProxyRequestEvent implements Serializable, Cloneable {
     private String body;
 
     private Boolean isBase64Encoded;
-
-    /**
-     * Class that represents Http Headers.
-     *
-     * Not using a standard map, because we need insensitive case.
-     */
-    public static class Headers<T> implements Map<String, T> {
-
-        // Headers are case insensitive (https://www.w3.org/Protocols/rfc2616/rfc2616-sec4.html#sec4.2)
-        private Map<String, T> map = new TreeMap<>(String.CASE_INSENSITIVE_ORDER);
-
-        @Override
-        public int size() {
-            return map.size();
-        }
-
-        @Override
-        public boolean isEmpty() {
-            return map.isEmpty();
-        }
-
-        @Override
-        public boolean containsKey(Object key) {
-            return map.containsKey(key);
-        }
-
-        @Override
-        public boolean containsValue(Object value) {
-            return map.containsValue(value);
-        }
-
-        @Override
-        public T get(Object key) {
-            return map.get(key);
-        }
-
-        @Override
-        public T put(String key, T value) {
-            return map.put(key, value);
-        }
-
-        @Override
-        public T remove(Object key) {
-            return map.remove(key);
-        }
-
-        @Override
-        public void putAll(Map<? extends String, ? extends T> m) {
-            map.putAll(m);
-        }
-
-        @Override
-        public void clear() {
-            map.clear();
-        }
-
-        @Override
-        public Set<String> keySet() {
-            return map.keySet();
-        }
-
-        @Override
-        public Collection<T> values() {
-            return map.values();
-        }
-
-        @Override
-        public Set<Entry<String, T>> entrySet() {
-            return map.entrySet();
-        }
-    }
 
     /**
      * class that represents proxy request context
@@ -1015,14 +946,19 @@ public class APIGatewayProxyRequestEvent implements Serializable, Cloneable {
      * @return The headers sent with the request
      */
     public Map<String, String> getHeaders() {
-        return headers.map;
+        return headers;
     }
 
     /**
      * @param headers The headers sent with the request
      */
     public void setHeaders(Map<String, String> headers) {
-        this.headers.putAll(headers);
+        if (this.headers == null && headers != null && !headers.isEmpty()) {
+            this.headers = new HttpHeaders<>();
+        }
+        if (headers != null && !headers.isEmpty()) {
+            this.headers.putAll(headers);
+        }
     }
 
     /**
@@ -1045,7 +981,12 @@ public class APIGatewayProxyRequestEvent implements Serializable, Cloneable {
      * @param multiValueHeaders The multi value headers sent with the request
      */
     public void setMultiValueHeaders(Map<String, List<String>> multiValueHeaders) {
-        this.multiValueHeaders.putAll(multiValueHeaders);
+        if (this.multiValueHeaders == null && multiValueHeaders != null && !multiValueHeaders.isEmpty()) {
+            this.multiValueHeaders = new HttpHeaders<>();
+        }
+        if (multiValueHeaders != null && !multiValueHeaders.isEmpty()) {
+            this.multiValueHeaders.putAll(multiValueHeaders);
+        }
     }
 
     /**
