+++ aws-lambda-java-events/src/main/java/com/amazonaws/services/lambda/runtime/events/APIGatewayProxyResponseEvent.java
@@ -1,5 +1,7 @@
 package com.amazonaws.services.lambda.runtime.events;
 
+import com.amazonaws.services.lambda.runtime.events.models.HttpHeaders;
+
 import java.io.Serializable;
 import java.util.List;
 import java.util.Map;
@@ -13,9 +15,9 @@ public class APIGatewayProxyResponseEvent implements Serializable, Cloneable {
     
     private Integer statusCode;
 
-    private Map<String, String> headers;
+    private HttpHeaders<String> headers;
 
-    private Map<String, List<String>> multiValueHeaders;
+    private HttpHeaders<List<String>> multiValueHeaders;
     
     private String body;
 
@@ -60,7 +62,12 @@ public class APIGatewayProxyResponseEvent implements Serializable, Cloneable {
      * @param headers The Http headers return in the response
      */
     public void setHeaders(Map<String, String> headers) {
-        this.headers = headers;
+        if (this.headers == null && headers != null && !headers.isEmpty()) {
+            this.headers = new HttpHeaders<>();
+        }
+        if (headers != null && !headers.isEmpty()) {
+            this.headers.putAll(headers);
+        }
     }
 
     /**
@@ -83,7 +90,12 @@ public class APIGatewayProxyResponseEvent implements Serializable, Cloneable {
      * @param multiValueHeaders the Http multi value headers to return in the response
      */
     public void setMultiValueHeaders(Map<String, List<String>> multiValueHeaders) {
-        this.multiValueHeaders = multiValueHeaders;
+        if (this.multiValueHeaders == null && multiValueHeaders != null && !multiValueHeaders.isEmpty()) {
+            this.multiValueHeaders = new HttpHeaders<>();
+        }
+        if (multiValueHeaders != null && !multiValueHeaders.isEmpty()) {
+            this.multiValueHeaders.putAll(multiValueHeaders);
+        }
     }
 
     /**
