+++ aws-lambda-java-events/src/main/java/com/amazonaws/services/lambda/runtime/events/ApplicationLoadBalancerRequestEvent.java
@@ -1,5 +1,6 @@
 package com.amazonaws.services.lambda.runtime.events;
 
+import com.amazonaws.services.lambda.runtime.events.models.HttpHeaders;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
@@ -40,8 +41,8 @@ public class ApplicationLoadBalancerRequestEvent implements Serializable, Clonea
     private String path;
     private Map<String, String> queryStringParameters;
     private Map<String, List<String>> multiValueQueryStringParameters;
-    private Map<String, String> headers;
-    private Map<String, List<String>> multiValueHeaders;
+    private HttpHeaders<String> headers;
+    private HttpHeaders<List<String>> multiValueHeaders;
     private String body;
     private boolean isBase64Encoded;
