+++ aws-lambda-java-tests/src/test/resources/apigw_http_event.json
@@ -8,8 +8,9 @@
     "cookie2"
   ],
   "headers": {
-    "Header1": "value1",
-    "Header2": "value1,value2"
+    "Header1": "Value1",
+    "Header2": "value1,value2",
+    "header1": "value1"
   },
   "queryStringParameters": {
     "parameter1": "value1,value2",
