+++ aws-lambda-java-tests/src/test/resources/elb_event.json
@@ -9,6 +9,7 @@
   "queryStringParameters": {},
   "headers": {
     "accept": "text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8",
+    "Accept": "application/json",
     "accept-encoding": "gzip",
     "accept-language": "en-US,en;q=0.5",
     "connection": "keep-alive",
